% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/rapgen.R
\docType{package}
\name{rapgen}
\alias{rapgen}
\alias{rapgen-package}
\title{rapgen: Multi-registry reports and results at Rapporteket}
\description{
An R package for functions and stuff that neither fits into 'rapbase' nor
 any registry spesiffic package. Initially set up for reports and results
 spanning across registries at Rapporteket
}
\section{rapgen functions}{

\code{\link{RegDelay}}
\code{\link{RegDelayData}}
\code{\link{RegDelayDataPeculiar}}
\code{\link{RegDelayReport}}
}

\section{rapgen reports}{

RegDelayReport.Rmd
}

