% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/RegDelayData.R, R/RegPulseData.R
\name{RegDelayData}
\alias{RegDelayData}
\title{Provide registration delay data from OpenQReg quality registries}
\format{Return a data frame with two variables:
\describe{
\item{year}{the year for the event (from db field \emph{HovedDato})}
\item{daysDiff}{the difference in days between event (from db field
\emph{HovedDato}) and the last time the registration form was saved
(from db field \emph{SistLagretDato})}
}}
\usage{
RegDelayData(registryName, startDate, endDate)

RegDelayData(registryName, startDate, endDate)
}
\arguments{
\item{registryName}{String Key/name of registry to get data from}

\item{startDate}{String Date defining the lower limit (endpoint included) of
data sample from the registry given as 'YYYY-MM-DD'}

\item{endDate}{String Date defining the upper limit (endpoint included) of
data sample from the registry given as 'YYYY-MM-DD'}

\item{registrationFormName}{String Name of the registry form where dates
are to be collected from}

\item{registryName}{String Key/name of registry to get data from}
}
\value{
regDelayData data frame

regPulseData Date Vector of dates
}
\description{
Provides delay between event (main date) and registration as a data frame

Provides registrarion dates from an OpenQReg register. The aim is to obtain
some sort of "pulse", that is how frequent data are being supplied (and NOT
how much)
}
\details{
For the query these conditions apply:
\describe{
\item{SkjemaStatus = 1}{use Only finished registrations}
\item{SkjemaNavn = registrationFormName}{use only registry form name given
by the input argument \emph{registrationFormName}. For surgical registries
this is usually the form "Operasjon".}
}

For the query these conditions apply:
\describe{
\item{Dates are collected from table \emph{SkjemaOversikt}, field
 \emph{OpprettetDato}}
\item{All registry forms (field \emph{SkjemaNavn}) are valid}
\item{All registrarion states (field \emph{SkjemaStatus}) are valid}
}
}
\seealso{
This function is used by \code{\link{RegDelay}}.

This function is used by \code{\link{RegDelay}}.
}

