% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/RegPulse.R
\name{RegPulse}
\alias{RegPulse}
\title{Provide registration "pulse" for OpenQReg quality registries}
\usage{
RegPulse(registryName, startDate, endDate, peculiarity = FALSE)
}
\arguments{
\item{peculiarity}{Logical TRUE for none-OpenQReg registries. False by
default}

\item{years}{integer vector with years for results and grouping}
}
\value{
data frame with registry name and "pulse" for each period
}
\description{
The aim is to obtain some sort of "pulse", that is during a given period how
frequent are data being supplied (and NOT how much). This pulse is provided
yearly, monthly and weekly periods measured by the ratio of 'empty' periods
wthin the total timeframe (sample intervall). \emph{E.g.} if over a
timeframe of 12 months registrations occur in 6 of them the montly pulse
becomes 0.5
}
\seealso{
Data to this function is provided by \code{\link{RegPulseData}}
and \code{\link{RegPulseDataPeculiar}}.
}

